tests/wpt/web-platform-tests/css/css-box/parsing/height-invalid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS basic box model: parsing height with invalid values</title>
<link rel="author" title="Eric Willigers" href="mailto:ericwilligers@chromium.org">
<link rel="help" href="https://drafts.csswg.org/css-box-3/#propdef-height">
<meta name="assert" content="height supports only the grammar '[<length> | <percentage>] | available | min-content | max-content | fit-content | complex | auto'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_invalid_value("height", "none");

test_invalid_value("height", "min-content available");
test_invalid_value("height", "max-content 10px");
test_invalid_value("height", "20% available");

test_invalid_value("height", "-10px");
test_invalid_value("height", "-20%");
test_invalid_value("height", "60");
test_invalid_value("height", "10px 20%");

test_invalid_value("height", "10px border-box");
test_invalid_value("height", "content-box 20%");
</script>
</body>
</html>
